ui fixed. viles changer fixed.
pressing on scummy will not close the app anymore but will take u to main menu.

added an option to go to the main menu from the pause menu.

added the E button in the ui to show the interact button.

i added a randomizer for the viles for a better puzzle.